redo available_scripts test from scratch
